+++ TeamCode/src/main/java/teamcode/mule/MuleRobot.kt
@@ -7,7 +7,7 @@ class MuleRobot() {
     val hardware = MuleHardware()
 
     val arm = MuleArm(hardware.armMotor)
-    val claw = Claw(hardware.clawServo, hardware.distanceSensor)
+    val claw = Claw(hardware.clawServo)
 
     init {
         arm.setPos(-67.7)
