+++ TeamCode/src/main/java/teamcode/v1/auto/AutoRobot.kt
@@ -20,8 +20,8 @@ class AutoRobot(startPose: Pose) {
     )
 
     val arm = Arm(hardware.armMotor, hardware.limitSwitch)
-    val claw = Claw(hardware.clawServo, hardware.distanceSensor)
-    val guide = Guide(hardware.guideServo)
+    val claw = Claw(hardware.clawServo)
+    val guide = Guide(hardware.guideServo, hardware.distanceSensor)
     val whacker = Whacker(hardware.whackerServo)
     val vision = Vision()
     val lift = Lift(
