+++ TeamCode/src/main/java/teamcode/v1/constants/Strategy.kt
@@ -1,0 +1,8 @@
+package teamcode.v1.constants
+
+enum class Strategy {
+    GROUND,
+    LOW,
+    MID,
+    HIGH,
+}
